% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow2.R
\name{bay}
\alias{bay}
\title{bay function}
\usage{
bay(k)
}
\arguments{
\item{k}{is a data set.can be a vector,matrix,data frame, table.}
}
\value{
a side by side bar plot object that shows the relationship between the
various variables
}
\description{
bay function
}
\examples{
bay()
}
