% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow2.R
\name{jael}
\alias{jael}
\title{Expected cell frequencies}
\usage{
jael(x)
}
\arguments{
\item{x}{is a data set.which can be a vector,matrix,data frame, table}
}
\value{
y a numeric value
}
\description{
Expected cell frequencies
}
\examples{
jael()
}
