% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow2.R
\name{mos}
\alias{mos}
\title{Mos function}
\usage{
mos(h)
}
\arguments{
\item{h}{is a data set.which can be a vector,matrix,data frame, table.}
}
\value{
a mosaic plot object of the contigency table
}
\description{
Mos function
}
\examples{
mos()
}
